feat: refactor the client
refactor: Update code examples to use `no_run` attribute

refactor: Update RPC client payload handling in girolle crate

This commit updates the `girolle` crate to include the `RpcCall` trait in the public API. This allows users to make RPC calls using the `RpcClient` struct. Additionally, the `RpcClient` struct now includes the `RpcCall` trait as a public dependency.

The changes in this commit improve the overall code structure and make it easier for users to interact with the RPC functionality provided by the `girolle` crate.

Closes #103

refactor: Update RPC client payload handling in girolle crate

feat: Add tokio dependency for async timing in rpc_client.rs

refactor: Update RPC client payload handling in girolle crate

refactor: Update RPC client payload handling in girolle crate

refactor: Create RpcCall struct in RpcClient

This commit adds a new function `create_rpc_call` to the `RpcClient` struct in `rpc_client.rs`. The function creates a new `RpcCall` struct for a given service name in the Nameko microservice. This allows users to make RPC calls using the `RpcClient` struct. The changes improve the overall code structure and make it easier for users to interact with the RPC functionality provided by the `girolle` crate.

refactor: Update sleep duration in simple_sender.rs and add Clone trait to Payload struct

refactor: Add 'sub' RPC method to GreetingService

This commit adds the 'sub' method to the 'GreetingService' class in the 'nameko_service.py' file. The 'sub' method performs subtraction of two numbers and returns the result. This addition enhances the functionality of the service by providing a new RPC method for clients to use.

Closes #103

refactor: Remove unused variable in RpcClient

This commit removes the unused variable `channel` in the `RpcClient` struct in `rpc_client.rs`. The variable was not being used and was causing unnecessary clutter in the code. Removing it improves code readability and maintainability.

Closes #103

feat: Add 'sub' RPC method to GreetingService

This commit adds the 'sub' method to the 'GreetingService' class in the 'nameko_service.py' file. The 'sub' method performs subtraction of two numbers and returns the result. This addition enhances the functionality of the service by providing a new RPC method for clients to use.

refactor: rename variable